Create Fibonacci48.java
The Fibonacci sequence is attributed to Italian mathematician Leonardo of Pisa. It is represented as a sequence of integers, starting with 0, 1 where every number after the first 2 is the sum of the two preceding numbers. The beginning of the Fibonacci sequence looks like...

0, 1, 1, 2, 3, 5, 8,13, 21,34

Your task is to write a program which renders the the number of the Fibonacci sequence. For example if given an  n value of 9, you are to return the 9th number of the sequence, which would be the number 21.

Input Format

The first line will contain a positive integer T denoting the number of test cases that follow. Each test case will have the following input:

- A single positive integer which represents the position in the sequence for which you are to determine the value.

Constraints

0 < T <= 90


Output Format

-For each test case, your program should output one line containing the sequence position and the value of the integer at that position in the Fibonacci sequence. Each line of output should be in the format:

POSITION = VALUE

Sample Input 0

7
1
5
8
11
13
21
40


Sample Output 0

1 = 0
5 = 3
8 = 13
11 = 55
13 = 144
21 = 6765
40 = 63245986